pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastmcp>=2.11.2",
    "google-genai>=1.29.0",
    "httpx>=0.24.0",
    "markdownify>=1.1.0",
    "mcp[cli]>=1.12.4",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
]
